Make it testable
